Issue/724: feature: New Skill rolling logic
