--- business/dice/ability-roll/attribute-and-skill-roll-schema/attribute-and-skill-roll-schema.mjs
@@ -1,268 +1,0 @@
-import { VISIBILITY_MODES } from "../../../../presentation/chat/visibility-modes.mjs";
-import ChoiceOption from "../../../../presentation/component/input-choice/choice-option.mjs";
-import DynamicInputDialog from "../../../../presentation/dialog/dynamic-input-dialog/dynamic-input-dialog.mjs";
-import { ACTOR_TYPES } from "../../../document/actor/actor-types.mjs";
-import TransientSkill from "../../../document/item/skill/transient-skill.mjs";
-import CharacterAttribute from "../../../ruleset/attribute/character-attribute.mjs";
-import { Sum, SumComponent } from "../../../ruleset/summed-data.mjs";
-import GameSystemUserSettings from "../../../setting/game-system-user-settings.mjs";
-import RollData from "../../roll-data.mjs";
-import { ROLL_DICE_MODIFIER_TYPES } from "../../roll-dice-modifier-types.mjs";
-import { RollSchema } from "../../roll-schema.mjs";
-import { SkillRollSchema } from "../skill-roll-schema.mjs";
-import AttributeAndSkillRollQueryData from "./attribute-and-skill-roll-query-data.mjs";
-import DynamicInputDefinition from "../../../../presentation/dialog/dynamic-input-dialog/dynamic-input-definition.mjs";
-import InputTextFieldViewModel from "../../../../presentation/component/input-textfield/input-textfield-viewmodel.mjs";
-import { ValidationUtil } from "../../../util/validation-utility.mjs";
-import InputDropDownViewModel from "../../../../presentation/component/input-choice/input-dropdown/input-dropdown-viewmodel.mjs";
-import InputNumberSpinnerViewModel from "../../../../presentation/component/input-number-spinner/input-number-spinner-viewmodel.mjs";
-import { StringUtil } from "../../../util/string-utility.mjs";
-import TransientBaseActor from "../../../document/actor/transient-base-actor.mjs";
-
-/**
- * Defines a schema for rolling dice to test a skill. 
- * 
- * This schema adds the level of only one base attribute to the level of the skill. 
- * The resulting sum is the number of dice to roll. 
- * 
- * @extends SkillRollSchema
- */
-export class AttributeAndSkillRollSchema extends SkillRollSchema {
-  /**
-   * @param {Object} args 
-   * @param {Number | undefined} args.dieFaces The number of faces on a die. 
-   * * default `6`
-   * @param {Number | undefined} args.hitThreshold Sets the lower bound of faces that are considered 
-   * hits. Any face turning up this number and numbers above, are considered hits. 
-   * * default `5`
-   * 
-   * @param {TransientBaseActor | undefined} args.owningDocumentOverride If set, will be treated as the Skill's parent, 
-   * even if it isn't. This is for use in synthetic Skill rolling. 
-   */
-  constructor(args = {}) {
-    super(args);
-
-    this._owningDocumentOverride = args.owningDocumentOverride;
-  }
-
-  /**
-   * @param {TransientSkill} document 
-   * @param {AttributeAndSkillRollQueryData} rollQueryData 
-   * 
-   * @override
-   */
-  async getRollData(document, rollQueryData) {
-    const diceComponents = this._getDiceComponents(document, rollQueryData.attribute);
-
-    return new RollData({
-      dieFaces: this.dieFaces,
-      hitThreshold: this.hitThreshold,
-      obFormula: rollQueryData.ob,
-      diceComponents: new Sum(diceComponents),
-      bonusDiceComponent: new SumComponent(RollSchema.BONUS_DICE_COMPONENT, "system.roll.bonusDice", rollQueryData.bonusDice),
-      hitModifier: 0,
-      compensationPoints: rollQueryData.compensationPoints,
-      rollModifier: rollQueryData.rollModifier,
-    });
-  }
-
-  /** @override */
-  getAvailableDiceComponents(document) {
-    const sums = [];
-
-    document.baseAttributes.forEach(baseAttribute => {
-      const owningDocument = this._owningDocumentOverride ?? document.owningDocument;
-      const characterAttribute = new CharacterAttribute(owningDocument.document, baseAttribute.name);
-      const attributeLevel = characterAttribute.modifiedLevel;
-      sums.push(new Sum([
-        new SumComponent("attribute", baseAttribute.localizableName, attributeLevel),
-        new SumComponent("skill", document.name, document.modifiedLevel),
-      ]));
-    });
-    sums.sort((a, b) => b.total - a.total);
-
-    return sums;
-  }
-
-  /** @override */
-  getAvailableDiceComponentExplanation(document) {
-    const sums = this.getAvailableDiceComponents(document);
-
-    const availableDiceStrings = sums.map(sum => {
-      const attributeComponent = sum.components.find(it => it.name === "attribute");
-      const skillComponent = sum.components.find(it => it.name === "skill");
-
-      return StringUtil.format2(game.i18n.localize("system.roll.availableDiceWithAttribute"), {
-        skillLevel: skillComponent.value,
-        skillName: game.i18n.localize(StringUtil.escapeHtml(skillComponent.localizableName)),
-        attributeLevel: attributeComponent.value,
-        attributeName: game.i18n.localize(StringUtil.escapeHtml(attributeComponent.localizableName)),
-        total: sum.total,
-      });
-    });
-    return game.i18n.localize("system.roll.availableDice") + "<br>" + availableDiceStrings.join("<br>");
-  }
-
-  /**
-   * @param {TransientSkill} document 
-   * @param {DynamicInputDialog} dialog 
-   * 
-   * @override
-   */
-  async _queryRollData(document, dialog) {
-    // Input definitions specific to dice pool rolls. 
-    const nameInputObstacle = "inputObstacle";
-    const nameInputAttribute = "inputAttribute";
-    const nameInputBonusDice = "inputBonusDice";
-    const nameInputCompensationPoints = "inputCompensationPoints";
-    const nameInputRollDiceModifier = "inputRollDiceModifier";
-
-    const owningDocument = this._owningDocumentOverride ?? document.owningDocument;
-    const actor = owningDocument.document;
-    const attributes = document.baseAttributes.map(attribute => new CharacterAttribute(actor, attribute.name));
-
-    const bestAttribute = this._getBestAttribute(attributes);
-    const baseAttributeChoices = attributes.map(attribute => new ChoiceOption({
-      value: attribute.name,
-      localizedValue: game.i18n.localize(attribute.localizableName),
-    }));
-
-    const availableDiceExplanation = this.getAvailableDiceComponentExplanation(document);
-
-    dialog.inputDefinitions.splice(0, 0, // Insert the following before the visibility drop down. 
-      new DynamicInputDefinition({
-        name: "diceCompositionLabel",
-        localizedLabel: `<p class="font-size-sm">${availableDiceExplanation}</p>`,
-        showFancyFont: false,
-      }),
-      new DynamicInputDefinition({
-        name: nameInputObstacle,
-        localizedLabel: game.i18n.localize("system.roll.obstacle.abbreviation"),
-        template: InputTextFieldViewModel.TEMPLATE,
-        viewModelFactory: (id, parent, overrides) => new InputTextFieldViewModel({
-          id: id,
-          parent: parent,
-          value: "0",
-          placeholder: game.i18n.localize("system.roll.obstacle.rollForPlaceholder"),
-          ...overrides,
-        }),
-        required: true,
-        validationFunc: (value) => { return ValidationUtil.isNotBlankOrUndefined(value); },
-      }),
-      new DynamicInputDefinition({
-        name: nameInputAttribute,
-        localizedLabel: game.i18n.localize("system.character.attribute.singular"),
-        template: InputDropDownViewModel.TEMPLATE,
-        viewModelFactory: (id, parent, overrides) => new InputDropDownViewModel({
-          id: id,
-          parent: parent,
-          isEditable: baseAttributeChoices.length > 1,
-          options: baseAttributeChoices,
-          value: baseAttributeChoices.find(choice => choice.value === bestAttribute.name),
-          ...overrides,
-        }),
-      }),
-      new DynamicInputDefinition({
-        name: nameInputBonusDice,
-        localizedLabel: game.i18n.localize("system.roll.bonusDice"),
-        template: InputNumberSpinnerViewModel.TEMPLATE,
-        viewModelFactory: (id, parent, overrides) => new InputNumberSpinnerViewModel({
-          id: id,
-          parent: parent,
-          value: 0,
-          ...overrides,
-        }),
-        required: true,
-        validationFunc: (value) => { return parseInt(value) !== NaN; },
-      }),
-      new DynamicInputDefinition({
-        name: nameInputCompensationPoints,
-        localizedLabel: game.i18n.localize("system.roll.compensationPoints"),
-        template: InputNumberSpinnerViewModel.TEMPLATE,
-        viewModelFactory: (id, parent, overrides) => new InputNumberSpinnerViewModel({
-          id: id,
-          parent: parent,
-          value: 0,
-          ...overrides,
-        }),
-        required: true,
-        validationFunc: (value) => { return parseInt(value) !== NaN; },
-      }),
-      new DynamicInputDefinition({
-        name: nameInputRollDiceModifier,
-        localizedLabel: game.i18n.localize("system.roll.diceModifier.plural"),
-        template: InputDropDownViewModel.TEMPLATE,
-        viewModelFactory: (id, parent, overrides) => new InputDropDownViewModel({
-          id: id,
-          parent: parent,
-          options: ROLL_DICE_MODIFIER_TYPES.asChoices(),
-          value: ROLL_DICE_MODIFIER_TYPES.asChoices().find(it => it.value === ROLL_DICE_MODIFIER_TYPES.NONE.name),
-          ...overrides,
-        }),
-      }),
-    );
-
-    const showReminders = new GameSystemUserSettings().get(GameSystemUserSettings.KEY_TOGGLE_REMINDERS);
-    const isPC = owningDocument.type === ACTOR_TYPES.PC;
-    if (showReminders && isPC) {
-      dialog.inputDefinitions.splice(1, 0, // Insert after the dice composition. 
-        new DynamicInputDefinition({
-          name: "forkReminderLabel",
-          localizedLabel: `<p>${game.i18n.localize("system.character.skill.forking.reminder.label")}</p>`,
-          showFancyFont: false,
-        }),
-      );
-    }
-
-    await dialog.renderAndAwait(true);
-    if (dialog.confirmed !== true) return undefined;
-
-    return new AttributeAndSkillRollQueryData({
-      attribute: attributes.find(attribute => attribute.name === dialog[nameInputAttribute].value),
-      ob: dialog[nameInputObstacle],
-      bonusDice: parseInt(dialog[nameInputBonusDice]),
-      compensationPoints: parseInt(dialog[nameInputCompensationPoints]),
-      rollModifier: ROLL_DICE_MODIFIER_TYPES.asArray().find(it => it.name === dialog[nameInputRollDiceModifier].value),
-      visbilityMode: VISIBILITY_MODES.asArray().find(it => it.name === dialog[this._nameInputVisibility].value),
-    });
-  }
-
-  /**
-   * Returns the dice components that comprise the sum of attribute and skill. 
-   * 
-   * @param {TransientSkill} document 
-   * @param {CharacterAttribute} characterAttribute
-   * 
-   * @returns {Array<SumComponent>}
-   * 
-   * @private
-   */
-  _getDiceComponents(document, characterAttribute) {
-    return [
-      new SumComponent(characterAttribute.name, characterAttribute.localizableName, characterAttribute.modifiedLevel),
-      new SumComponent(document.name, document.name, document.modifiedLevel),
-    ];
-  }
-
-  /**
-   * Returns the character attribute with the highest modified level. 
-   * 
-   * @param {Array<CharacterAttribute>} attributes 
-   * 
-   * @returns {CharacterAttribute}
-   * 
-   * @private
-   */
-  _getBestAttribute(attributes) {
-    let best = attributes[0];
-
-    attributes.forEach(attribute => {
-      if (attribute.modifiedLevel > best.modifiedLevel) {
-        best = attribute;
-      }
-    });
-
-    return best;
-  }
-
-}
