+++ business/ruleset/attribute/attribute-types.mjs
@@ -1,0 +1,11 @@
+/**
+ * A global constant for the Attribute classifications. 
+ * 
+ * @constant
+ */
+export const ATTRIBUTE_TYPES = {
+  SECONDARY: "secondary",
+  CORE: "core",
+  FAVORED: "favored",
+  PENALIZED: "penalized",
+};
