+++ business/ruleset/ruleset-explainer.mjs
@@ -91,7 +91,8 @@ export default class RulesetExplainer {
    */
   getExplanationForAttributeAdvancement(attribute) {
     return StringUtil.format2(game.i18n.localize("system.character.advancement.experiencePoint.requiredForAdvancement"), {
-      xp: new Ruleset().getAttributeAdvancementRequirements(attribute.level),
+      xp: new Ruleset().getAttributeAdvancementRequirements(attribute),
+      type: game.i18n.localize(`system.character.attribute.type.${attribute.type}`),
     });
   }
